Added XML docs and logging.
And moved `IVirtualizingListSource`  to GitHub.App.
