+++ src/GitHub.App/Collections/VirtualizingListCollectionView.cs
@@ -6,16 +6,29 @@ using System.Windows.Data;
 
 namespace GitHub.Collections
 {
+    /// <summary>
+    /// A <see cref="CollectionView"/> that adds filtering to a <see cref="VirtualizingList{T}"/>.
+    /// </summary>
+    /// <typeparam name="T">The item type.</typeparam>
     public class VirtualizingListCollectionView<T> : CollectionView, IList
     {
         List<int> filtered;
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="VirtualizingListCollectionView{T}"/> class.
+        /// </summary>
+        /// <param name="inner">The inner virtualizing list.</param>
         public VirtualizingListCollectionView(VirtualizingList<T> inner)
             : base(inner)
         {
         }
 
+        /// <summary>
+        /// Gets the count of the filtered items.
+        /// </summary>
         public override int Count => filtered?.Count ?? Inner.Count;
+
+        /// <inheritdoc/>
         public override bool IsEmpty => Count == 0;
 
         bool IList.IsReadOnly => true;
@@ -23,6 +36,9 @@ namespace GitHub.Collections
         object ICollection.SyncRoot => null;
         bool ICollection.IsSynchronized => false;
 
+        /// <summary>
+        /// Gets the inner virtualizing list.
+        /// </summary>
         protected VirtualizingList<T> Inner => (VirtualizingList<T>)SourceCollection;
 
         object IList.this[int index]
@@ -31,6 +47,7 @@ namespace GitHub.Collections
             set { throw new NotImplementedException(); }
         }
 
+        /// <inheritdoc/>
         public override object GetItemAt(int index)
         {
             if (filtered == null)
@@ -43,14 +60,14 @@ namespace GitHub.Collections
             }
         }
 
-        int IList.Add(object value) { throw new NotSupportedException(); }
-        bool IList.Contains(object value) { throw new NotImplementedException(); }
-        void IList.Clear() { throw new NotSupportedException(); }
-        int IList.IndexOf(object value) { throw new NotImplementedException(); }
-        void IList.Insert(int index, object value) { throw new NotSupportedException(); }
-        void IList.Remove(object value) { throw new NotSupportedException(); }
-        void IList.RemoveAt(int index) { throw new NotSupportedException(); }
-        void ICollection.CopyTo(Array array, int index) { throw new NotImplementedException(); }
+        int IList.Add(object value) => throw new NotSupportedException();
+        bool IList.Contains(object value) => throw new NotImplementedException();
+        void IList.Clear() => throw new NotSupportedException();
+        int IList.IndexOf(object value) => throw new NotImplementedException();
+        void IList.Insert(int index, object value) => throw new NotSupportedException();
+        void IList.Remove(object value) => throw new NotSupportedException();
+        void IList.RemoveAt(int index) => throw new NotSupportedException();
+        void ICollection.CopyTo(Array array, int index) => throw new NotImplementedException();
 
         protected override void RefreshOverride()
         {
